feature(#8): add basic routing with @reach/router
The reason for choosing to use @reach/router instead of react-router is
due to its minimalistic approach and simpler API. The latter has a lot
of features that we'll never use, e.g. HashRouter, MemoryRouter, and
StaticRouter, and requires a lot more code to achieve the same as can be
achieved with @reach/router — for typical usages.