bugfix: override default typing of ThemeConsumer
Ideally, we would be able to directly use the ThemeProvider and
ThemeConsumer components exported from the `styled-components` package.
However, both components are, unfortunately, hard-coded with `any` as
the type argument of `Consumer`; as such, we would end up losing type
safety if we use the components directly.

To keep our code type-safe, we need to override the default typing of
the `ThemeConsumer` component. An unfortunate side-effect of this
choice, however, is that every developer will need to remember to import
`ThemeConsumer` from the correct package to get type assistance/safety –
should they choose to use it instead of the `useTheme` hook, which is
highly unlikely.