feature: add CSS fade-in animation
A fade-in animation has been added to all HTML elements upon
first-render. The exact implementation might be controversial, but it
allows us to enforce the animation everywhere, i.e. the animation is
ubiquitous, while encapsulating its definition and implementation -
components do not know that they will have a fade-in animation on mount
nor how the animation is applied to them.

The only problem with the implementation in this commit is that the
fade-in animation will be overwritten if an element has its own
animation. This is usually not a problem, but it will be for components
that contain elements with their own animation without also having some
container elements wrapping the former. In this specific case, the
component's animated elements will have no fade-in animation upon
first-render – as their own animation overrides it and there are no
container elements to be rendered alongside them to cascade the
animation.

A concrete example of the problem described above can be seen with the
`Loader` component, whose styling had to be slightly modified in this
commit. The `Loader` component only contains a single element that has
its own animation. As such, when the component mounts, only the single
element's animation will be played - since it overrides the fade-in
animation and there are no container elements cascading the fade-in
animation to it.

There are multiple ways to "fix" the problem regarding "ubiquitous"
animations being overridden, but only one approach is able to maintain
encapsulation of the animations' implementation. The solution is to view
the ubiquitous animations as being a part of the application's theme. In
that case, components will only need to know that they must account for
some abstract ubiquitous animations when defining their own animation
through the theme's interface - like how they are aware of the theme's
typography and color palette when defining font sizes and colors,
respectively. Note that the solution, however, fails to encapsulate the
ubiquitous animations' definition.

Finally, for comparison, we describe a "solution" were the encapsulation
of the ubiquitous animations' implementation is not maintained. Imagine
that we require every element that has its own animation to be wrapped
in at least one container element within the same component, e.g. the
single element in the `Loader` component must be wrapped in a `div`.
This "solution" works, but it forces every component to carefully
structure their elements, potentially in round-about ways, such as to
avoid breaking any ubiquitous animations. To achieve this, each
component must know how the ubiquitous animations' are implemented to
know how they might be affected - thus breaking the encapsulation of the
animations' implementation.