bugfix(#42): grid and card width in Landing page
Previously, the grid didn't really behave as a grid and the cards would
be huge on a 4k monitor. Both of these issues are fixed by this commit
by using CSS grid instead of flexbox and by setting a max-width,
respectively.

Note that the change might break the app on IE and older browsers,
however, it should be fine as this app is only meant for internal usage.

The motivation for changing from flexbox to CSS grid is because the
latter is better at handling grid layouts. The only reason for why CSS
grid was not used initially was due to an outdated assumption that it
does not have great browser support - which is no longer true according
to https://caniuse.com/?search=css-grid.