bugfix(routing): add scroll to top hack
Due to accessibility reasons, reach-router will focus on the destination
page's content after a navigation. However, a side-effect of the focus
operation is to scroll to the focused content, which is an unexpected
behavior for users relying on visuals. To fix this issue, without
breaking the accessibility functionality, we must manually scroll to the
top of the page after each navigation.

For more details, see:
- https://github.com/reach/router/issues/242; and
- https://stackoverflow.com/a/56996986.