Change authConfig to set apiScope differently for Production and Development
The apiScope has to be set differently to work in Production and Development.
It will now check if the apiClientId environment variable is set to apply
production apiScope.

To be able to run the project locally you have to create a file in scripts/ called:
generate-local-secrets-env.sh or generate-local-secrets-env.ps1 and they need to export
the enviroment variables MAD_LEARNING_WEB_CLIENT_ID and MAD_LEARNING_TENANT_ID