Add eventStatus to form, event page and event list
Add SelectField in create event form to set eventStatus to either "Draft" or
"Published". eventStatus can be changed in the update event form aswell.
Events on the front page will be labelled if the eventStatus is set to
"Draft". When viewing a single event that belongs to you, the event will
be marked with a label at the top with either "Draft" or "Published".