Add component for list of hyperlinks on EventPage
The new component requires the field resourceLink to be set in the API
as an array containing {"name": "", "url": ""} objects. The provided
links are displayed above the participants on the EventPage. The links
are displayed with the url's icon using Clearbit logo API. If a logo
is not found a default image is displayed instead. This commit does not
include any way to add links to an event.